chore: move federation schemas into dedicated mod
